requirements.txt
absl-py
aiohttp
aiosignal
amqp
antlr4-python3-runtime
asgiref
asttokens
astunparse
async-timeout
attrs
bcrypt
billiard
boto3
botocore
cachetools
celery
# certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
contourpy
coverage
coveralls
cryptography
cycler
Cython
databases
dataclass-wizard
debugpy
Django
django-cors-headers
djangorestframework
docopt
docx2txt
exceptiongroup
executing
filelock
Flask
Flask-Cors
flatbuffers
fonttools
frozenlist
fsspec
future
gast
gensim
gluonnlp
google-auth
google-auth-oauthlib
google-pasta
graphviz
greenlet
grpcio
h5py
hanja
huggingface-hub
hydra-core
idna
image
importlib-metadata
importlib-resources
iniconfig
install
itsdangerous
jamo
JayDeBeApi
Jinja2
jmespath
joblib
JPype1
JPype1-py3
jsonlines
keras
Keras-Preprocessing
kiwisolver
# -e git+https://github.com/XAI-AI-LAB/NLP_INTEGRATE.git@b776f079b84c42d9045f23e4555553eb30daef21gg=kobert_tokenizer&subdirectory=NER/src/kobert-tokenizer/kobert_hf
kombu
konlpy
lexrankr
libclang
littleutils
lxml
Markdown
MarkupSafe
matplotlib
mecab-python
mecab-python3
multidict
mxnet
networkx
# numpy
nvidia-cublas-cu11
nvidia-cuda-nvrtc-cu11
nvidia-cuda-runtime-cu11
nvidia-cudnn-cu11
nvidia-ml-py3
oauthlib
olefile
omegaconf
opencv-python
opt-einsum
packaging
pandas
pdfminer.six
pdfplumber
pdftotext
# Pillow
pluggy
prompt-toolkit
protobuf
py
pyasn1
pyasn1-modules
pycparser
pyDeprecate
pykospacing @ git+https://github.com/haven-jeon/PyKoSpacing.git@e1979f7eeb4fcf1a456521152713482517fba627
PyMySQL
pyparsing
pytest
pytest-cov
python-dateutil
pytorch-crf @ git+https://github.com/kmkurn/pytorch-crf@e843718540512eb04d9b756fca04a0915affe175
pytorch-lightning
pytz
PyYAML
regex
requests
requests-oauthlib
rsa
s3transfer
sacremoses
scikit-learn
scipy
seaborn
sentencepiece
six
sklearn
smart-open
sorcery
SQLAlchemy
sqlparse
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tokenizers
tomli
# torch
# torchaudio
torchmetrics
# torchvision
tqdm
transformers
# typing_extensions
urllib3
vine
Wand
wcwidth
Werkzeug
wordcloud
wrapt
yarl
zipp
